Fx codefactor issues
