+++ tests/DucksNet.UnitTests/ResultTests.cs
@@ -3,8 +3,6 @@ using FluentAssertions;
 
 using DucksNet.SharedKernel.Utils;
 
-
-
 public class ResultTests 
 {
     [Fact]
